[Major] Fixed bug on extraction (no big impact)
There was an edge case with the heterozygous variant extraction.
It would either extract a few unnecessary het variants, this has
no impact, since they will not be rephased, or fail to extract
between 1 and 2 het variants at the end if the samle had less than
the fifo size het variants. This is a very rare case on real files
and the number of missed variants will be small. So the impact of
the bug is not that high.

This is now fixed and extraction works as intended.