[Major] Fixed bug on extraction (minor impact)
The FIFO extraction logic was still not perfect, therefore
it has been simplified to reduce the edge case possibilities,
it now extracts the correct het variants. The impact of this
bug was not that big because it was only a few edge cases
where a few extra variants were extracted.